Created folders to better organize the code into components, utils, events. Created renderToDOM.js and main.js (startApp). Got the google sign in button to render. Need to configure my firebase backend so that I can interact with the api.
