+++ utils/renderToDOM.js
@@ -1,0 +1,6 @@
+const renderToDOM = (divID, domString) => {
+  const selectedDiv = document.querySelector(divID);
+  selectedDiv.innerHTML = domString;
+};
+
+export default renderToDOM;
